--- onnx_chainer/testing/__init__.py
@@ -1,2 +1,0 @@
-from onnx_chainer.testing import test_mxnet  # NOQA
-from onnx_chainer.testing import test_onnxruntime  # NOQA
