+++ tests/functions_tests/test_activations.py
@@ -4,7 +4,7 @@ import chainer.links as L
 from chainer import testing
 
 from onnx_chainer.testing import input_generator
-from tests.helper import ONNXModelTest
+from onnx_chainer.testing.helper import ONNXModelTest
 
 
 @testing.parameterize(
