+++ tests/functions_tests/test_loss.py
@@ -4,7 +4,7 @@ import chainer
 from chainer import testing
 import numpy as np
 
-from tests.helper import ONNXModelTest
+from onnx_chainer.testing.helper import ONNXModelTest
 
 
 @testing.parameterize(
