+++ tests/functions_tests/test_maths.py
@@ -6,7 +6,7 @@ import numpy as np
 import pytest
 
 from onnx_chainer.testing import input_generator
-from tests.helper import ONNXModelTest
+from onnx_chainer.testing.helper import ONNXModelTest
 
 
 @testing.parameterize(
