+++ tests/functions_tests/test_noises.py
@@ -3,7 +3,7 @@ import chainer.functions as F
 from chainer import testing
 import numpy as np
 
-from tests.helper import ONNXModelTest
+from onnx_chainer.testing.helper import ONNXModelTest
 
 
 @testing.parameterize(
